Fix: Implicit conversion from 'unsigned long' to 'double'

+++ src/helpers/config_loader.hpp
@@ -268,7 +268,7 @@ inline bool config_loader_t::parse_bytes(std::string const& str, size_t& bytes) 
 
     if (!ss.eof())
         return false;
-    if (std::isnan(number) || number > std::numeric_limits<size_t>::max())
+    if (std::isnan(number) || number > static_cast<double>(std::numeric_limits<size_t>::max()))
         return false;
 
     bytes = static_cast<size_t>(number);
